fix(intake-forms): widen questionnaire_response.encounter to bigint
int(11) silently overflows for OpenEMR's modern bigint encounter ids
(e.g. 900100000006 saturates to 2147483647). Match forms.encounter and
form_encounter.encounter which already use bigint(20). The bug is
pre-existing OpenEMR but surfaced during intake-forms smoke testing.

Only questionnaire_response.encounter needed widening; the sibling table
form_questionnaire_assessments has no encounter column, and
form_upload_intake_form.encounter was already declared BIGINT.

+++ sql/database.sql
@@ -14345,7 +14345,7 @@ CREATE TABLE `questionnaire_response` (
   `questionnaire_id` varchar(255) DEFAULT NULL COMMENT 'Id for questionnaire content. String version of UUID',
   `questionnaire_name` varchar(255) DEFAULT NULL,
   `patient_id` int(11) DEFAULT NULL,
-  `encounter` int(11) DEFAULT NULL COMMENT 'May or may not be associated with an encounter',
+  `encounter` bigint(20) DEFAULT NULL COMMENT 'May or may not be associated with an encounter',
   `audit_user_id` int(11) DEFAULT NULL,
   `creator_user_id` int(11) DEFAULT NULL COMMENT 'user id if answers are provider',
   `create_time` datetime DEFAULT current_timestamp(),
